runtime precision
- increase capture to nano instead of secs because we can have quite some infos which should be sortet well

+++ storage/models/protocol_entry.go
@@ -13,5 +13,5 @@ type ProtocolEntry struct {
 	Level          string
 	Message        string
 	Data           string
-	Runtime        int64
+	Runtime        float64
 }
